+++ src/extend.js
@@ -1,0 +1,7 @@
+import { merge, curry } from 'ramda'
+
+import shape from './shape'
+
+const extend = (schema, input) => merge(input, shape(schema, input))
+
+export default curry(extend)