+++ CHANGELOG.md
@@ -1,9 +1,24 @@
 # Changelog
 
+## 0.2.1
+#### Changes
+- Add extend export
+- Refactor shape to be curried
+- add build and publish script
+#### Author
+Quirkles
+
+## 0.1.1
+#### Changes
+- Add getOr export
+#### Author
+Quirkles
+
 ## 0.1.0
 #### Changes
 - Add optional callback
 #### Author
+Quirkles
 
 ## 0.0.6
 #### Changes